requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-box
pyyaml
ensure
psycopg2-binary
sqlalchemy
scikit-learn